+++ client/app/center/messages/services/renderers/file.comment.renderer.service.js
@@ -9,7 +9,7 @@
     .service('FileCommentRenderer', FileCommentRenderer);
 
   /* @ngInject */
-  function FileCommentRenderer($filter, MessageCollection, RendererUtil) {
+  function FileCommentRenderer($filter, MessageCollection, RendererUtil, publicService) {
     var _templateTitle = '';
     var _template = '';
 
@@ -42,15 +42,17 @@
 
       return template({
         css: {
+          unshared: publicService.isFileUnshared(msg, true) ? 'unshared' : '',
           star: RendererUtil.getStarCssClass(msg),
           wrapper: isTitle ? 'comment-title' : 'comment-continue',
           archived: isArchived ? 'archived-file-with-comment' : '',
           disabledMember: RendererUtil.getDisabledMemberCssClass(msg)
         },
         file: {
+          hasPermission: publicService.hasFilePermission(msg, true),
           icon: $filter('fileIcon')(content),
           title: $filter('fileTitle')(content),
-          imageUrl: _getImageUrl(msg),
+          imageUrl: $filter('getPreview')(content, 'small'),
           hasPreview: $filter('hasPreview')(content)
         },
         hasStar: RendererUtil.hasStar(msg),
@@ -62,18 +64,6 @@
         },
         msg: msg
       });
-    }
-
-    /**
-     * image url 을 반환한다.
-     * @param {object} msg
-     * @returns {*}
-     * @private
-     */
-    function _getImageUrl(msg) {
-      var content = msg.feedback.content;
-      var url = content && content.extraInfo && content.extraInfo.smallThumbnailUrl;
-      return $filter('getFileUrl')(url);
     }
   }
 })();