+++ client/app/center/messages/services/renderers/file.renderer.service.js
@@ -10,7 +10,7 @@
 
   /* @ngInject */
   function FileRenderer($filter, modalHelper, MessageCollection, RendererUtil, Loading, centerService, EntityMapManager,
-                        memberService, fileAPIservice, jndPubSub, AnalyticsHelper, currentSessionHelper) {
+                        memberService, fileAPIservice, jndPubSub, AnalyticsHelper, currentSessionHelper, publicService) {
     var _template = '';
 
     this.render = render;
@@ -140,47 +140,6 @@
     }
 
     /**
-     * file 접근 권한이 존재하는지 여부를 반환 (MK의 server 로직을 그대로 가져 옴.)
-     * @param {object} msg
-     * @returns {boolean}
-     * @private
-     */
-    function _hasPermission(msg) {
-      var message = msg.message;
-      var member = memberService.getMember();
-      var commonEntities = _.intersection(_.map(member.u_messageMarkers, 'entityId'), message.shareEntities);
-
-      // 파일이 전체 공개인 경우
-      if (message.permission % 10 > 0) {
-        return true;
-      }
-
-      // 파일 작성자가 본인인 경우
-      if (_isFileOwner(msg)) {
-        return true;
-      }
-
-      // 파일이 접근 가능한 entity에 공유된 경우
-      if (commonEntities.length > 0) {
-        return true;
-      }
-
-      return false;
-    }
-
-    /**
-     * unshared 된 file 인지 여부를 반환한다
-     * @param {object} msg
-     * @returns {boolean}
-     * @private
-     */
-    function _isUnshared(msg) {
-      var message = msg.message;
-      var currentEntityId = currentSessionHelper.getCurrentEntityId(true);
-      return message.shareEntities.indexOf(currentEntityId) === -1;
-    }
-
-    /**
      * index 에 해당하는 메세지를 랜더링한다.
      * @param {number} index
      * @returns {*}
@@ -190,15 +149,19 @@
       var content = msg.message.content;
       
       var isArchived = (msg.message.status === 'archived');
-      var hasPreview = $filter('hasPreview')(content);
       var icon = $filter('fileIcon')(content);
+  
+      var isUnshared = publicService.isFileUnshared(msg);
+      var hasPermission = publicService.hasFilePermission(msg);
+      var isMustPreview = $filter('mustPreview')(content);
 
       return _template({
         html: {
           loading: Loading.getTemplate()
         },
         css: {
-          unshared: _isUnshared(msg) ? 'unshared' : '',
+          unshared: isUnshared ? 'unshared' : '',
+          imageUnshared: (isMustPreview && isUnshared) ? 'image-unshare' : '',
           wrapper: isArchived ? ' archived-file': '',
           star: RendererUtil.getStarCssClass(msg),
           disabledMember: RendererUtil.getDisabledMemberCssClass(msg)
@@ -207,12 +170,13 @@
           download: _getFileDownloadAttrs(msg)
         },
         file: {
-          hasPermission: _hasPermission(msg),
+          unshared: isUnshared,
+          hasPermission: hasPermission,
           icon: icon,
           isImageIcon: icon === 'img',
-          mustPreview: $filter('mustPreview')(content),
-          hasPreview: hasPreview,
-          imageUrl: _getMediumThumbnailUrl(content, hasPreview),
+          mustPreview: isMustPreview,
+          hasPreview: $filter('hasPreview')(content),
+          imageUrl: $filter('getPreview')(content, 'medium'),
           title: $filter('fileTitle')(content),
           type: $filter('fileType')(content),
           size: $filter('bytes')(content.size),
@@ -250,19 +214,6 @@
      */
     function _getFeedbackContent(msg) {
       return centerService.isCommentType(msg.message.contentType) ? msg.feedback.content : msg.message.content;
-    }
-
-    /**
-     * image url 을 반환한다.
-     * @param {object} msg
-     * @returns {*}
-     * @private
-     */
-    function _getMediumThumbnailUrl(content, hasPreview) {
-      //var content = _getFeedbackContent(msg);
-      //var hasPreview = $filter('hasPreview')(content);
-
-      return hasPreview ? $filter('getFileUrl')(content.extraInfo.mediumThumbnailUrl) : '';
     }
 
     /**